Ensure ElementNode scopes are checked properly.
The changes in 8c2aaa7 fixed a number of issues with the way block
params were tracked for elements (and blocks) in the system. By changing
the scoping mechanism to only push the frame when we entered the
ElementNode's children the original bug was fixed, however the
`no-unused-block-params` rule was accidentally left checking for unused
locals directly in `ElementNode`'s `exit` handler. Unfortunately, by
that time the scope frame for the `ElementNode` had already been cleared
so we were checking the _outer_ scope frame's locals.

This fixes the `no-unused-block-params` rule by bringing it in line with
the mechansim that we use for pushing and popping scope frames.