Remove tests for no-trailing-dot-in-path-expression.
This error is actually created from `@glimmer/syntax` now (if you check
the rule's visitor it is `return {}`) and is only kept around here to
avoid a breaking change for folks that have it enabled in their config,
there is no need for us to test _how_ glimmer itself works.

Keeping the file around (as an empty file) until we can delete the rule
itself (this makes our "all rules must have a test" rule happy).