Refactor printers.
Refactor the argument passed in to each printer into the following
interface:

```ts
interface Message {
  rule: string;
  severity: 0 | 1 | 2;
  filePath?: string;
  moduleId?: string;
  message: string;
  line: number;
  column: number;
  source: string;
  isFixable?: boolean;
}

interface FileResult {
  filePath: string;
  messages: Message[];
  errorCount: number;
  warningCount: number;
  fixableErrorCount: number;
  fixableWarningCount: number;
}

interface TemplateLintResult {
  files: {
    [filePath: string]: FileResult;
  }

  errorCount: number;
  warningCount: number;
  fixableErrorCount: number;
  fixableWarningCount: number;
}
```

This simplifies the work the printers need to do, and also makes the
logic a bit less "hairy" in the bin script. This also ensures that the
relative file path is printed (vs printing the absolute path).